close the form on submit

+++ src/containers/Profile/ProfileForm.js
@@ -29,9 +29,11 @@ export class ProfileForm extends Component {
     handleDayChange(day){
       this.setState({dob: new Date(day).toLocaleDateString()})
     }
+
     handleGenderChange(vals){
       this.setState({gender: vals})
     }
+
     onStateSelect = (event, state) => {
     // event {SyntheticEvent<HTMLSelectElement>} - React HTML event
     // state {Object} - Object representing the state
@@ -39,6 +41,7 @@ export class ProfileForm extends Component {
     // state.abbreviation {string} - The two letter abbreviation of the states name
       this.setState({state: state.name})
     }
+
     render() {
         let delProfile = this.props.deleteProfile ? (<Button color='danger' className='btn-block wide-text profile-form-save-btn' size='lg' onClick={ () => this.deleteProfile() }>Delete</Button>) : ""
         return (
@@ -108,27 +111,28 @@ export class ProfileForm extends Component {
     }
 
     submitProfile() {
-      if(!this.state.id || this.state.id==""){
-        this.state.name =  this.state.first_name +" "+this.state.last_name;
-      }
-      this.props.saveProfile(this.state);
+        if (!this.state.id || this.state.id=="") {
+            this.state.name =  this.state.first_name +" "+this.state.last_name;
+        }
+        this.props.saveProfile(this.state);
+        this.props.cancel();
     }
 
-    deleteProfile(){
-      confirmAlert({
-      title: 'Confirm Profile Delete',
-      message: 'Are you sure you want to delete this profile?',
-      buttons: [
-        {
-          label: 'Yes',
-          onClick: () => this.props.deleteProfile(this.state.id)
-        },
-        {
-          label: 'No',
-          onClick: () => {}
-        }
-      ]
-    })
+    deleteProfile() {
+        confirmAlert({
+            title: 'Confirm Profile Delete',
+            message: 'Are you sure you want to delete this profile?',
+            buttons: [
+                {
+                label: 'Yes',
+                onClick: () => this.props.deleteProfile(this.state.id)
+                },
+                {
+                label: 'No',
+                onClick: () => {}
+                }
+            ]
+        })
     }
 }
